Split out TCPConnectionPool and add metrics
